requirements.txt
# Core ML
torch>=2.0.0,<3.0.0
transformers==4.42.0
transformer-lens>=1.0.0,<2.0.0
sae-lens>=0.5.0,<1.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
scipy>=1.10.0,<2.0.0
pytest

# HuggingFace
huggingface-hub>=0.20.0,<1.0.0
sentence-transformers>=2.2.0,<3.0.0

# Data & Visualization
pandas>=2.0.0,<3.0.0
matplotlib>=3.7.0,<4.0.0

# CLI & Config
click>=8.0.0,<9.0.0
pydantic>=2.0.0,<3.0.0

# Utilities
tqdm>=4.65.0,<5.0.0
absl-py>=1.0.0,<3.0.0
termcolor>=2.0.0,<3.0.0
reactivex>=4.0.0,<5.0.0
typing_extensions>=4.0.0,<5.0.0
requests>=2.28.0,<3.0.0

typeguard
accelerate==0.33.0

# DeepEval for LLM-based deception evaluation
deepeval>=0.21.0
openai>=1.0.0
python-dotenv>=1.0.0